+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.1.3] 2019-07-29.
+
+* Problem #2 with nipRadius.
+
+## [0.1.2] 2019-07-29.
+
+* Problem with nipRadius.
+
+## [0.1.1] 2019-07-29.
+
+* Small features.
+
 ## [0.1.0] 2019-07-29.
 
 * Ready for using from https://pub.dev/.